embed project into monorepo context as long as PR is open
Using lein-v/monolith allows projects from the outside to reference
gorillalabs/ring-anti-forgery in a way which is consistent to the other
lib projects when used as one project in the broader monorepo.

When run in the context of the monolith project, the version info from
the surrounding git is taken, when used standalone the version info
from the ring/ring-anti-forgery repo is taken.